requirements.txt
accelerate==1.1.1
datasets==3.1.0
deepspeed==0.15.4
flash-attn==2.7.0.post2
hydra-core==1.3.2
peft==0.13.2
torch==2.5.1
transformers==4.46.3